Dejando el Rank de escuelas en BS4.
